news-fact-checker/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
pymongo==4.6.0
langchain-perplexity==0.1.0
langchain-core==0.1.0
requests>=2.28.1
beautifulsoup4>=4.11.1
aiohttp>=3.8.3
motor>=3.3.0
pytest>=7.0.0,<8.0.0
mongomock>=4.1.0,<5.0.0
